refactor file imports into shared parent class
