+++ app/services/samples/batch_file_import_service.rb
@@ -4,111 +4,21 @@ require 'roo'
 
 module Samples
   # Service used to batch create samples via a file
-  class BatchFileImportService < BaseService # rubocop:disable Metrics/ClassLength
-    SampleFileImportError = Class.new(StandardError)
-
+  class BatchFileImportService < BaseSpreadsheetImportService
     def initialize(namespace, user = nil, blob_id = nil, params = {})
-      super(user, params)
-      @namespace = namespace
-      @file = ActiveStorage::Blob.find(blob_id)
       @sample_name_column = params[:sample_name_column]
       @project_puid_column = params[:project_puid_column]
       @sample_description_column = params[:sample_description_column]
-      @spreadsheet = nil
-      @headers = nil
-      @temp_import_file = Tempfile.new
+      required_headers = [@sample_name_column, @project_puid_column]
+      super(namespace, user, blob_id, required_headers, params)
     end
 
     def execute
       authorize! @namespace, to: :update_sample_metadata?
-
-      validate_sample_name_column
-
-      validate_file
-
-      perform_file_import
-    rescue Samples::BatchFileImportService::SampleFileImportError => e
-      @namespace.errors.add(:base, e.message)
-      {}
+      super
     end
 
-    private
-
-    def validate_sample_name_column
-      return unless @sample_name_column.nil?
-
-      raise SampleFileImportError,
-            I18n.t('services.samples.metadata.import_file.empty_sample_id_column') # TODO: refactor out
-    end
-
-    def validate_file_extension
-      file_extension = File.extname(@file.filename.to_s).downcase
-
-      return file_extension if %w[.csv .tsv .xls .xlsx].include?(file_extension)
-
-      raise SampleFileImportError,
-            I18n.t('services.samples.metadata.import_file.invalid_file_extension') # TODO: refactor out
-    end
-
-    def validate_file_headers
-      duplicate_headers = @headers.find_all { |header| @headers.count(header) > 1 }.uniq
-      unless duplicate_headers.empty?
-        raise SampleFileImportError,
-              I18n.t('services.sammple.batch_import.duplicate_column_names')
-      end
-
-      unless @headers.include?(@sample_name_column)
-        raise SampleFileImportError,
-              I18n.t('services.samples.batch_import.missing_header',
-                     header_title: @sample_name_column, header_desc: 'sample name')
-      end
-
-      unless @headers.include?(@project_puid_column) # rubocop:disable Style/GuardClause
-        raise SampleFileImportError,
-              I18n.t('services.samples.batch_import.missing_header',
-                     header_title: @project_puid_column, header_desc: 'project puid')
-      end
-    end
-
-    def validate_file_rows
-      # Should have at least 2 rows
-      first_row = @spreadsheet.row(2)
-      return unless first_row.compact.empty?
-
-      raise SampleFileImportError,
-            I18n.t('services.samples.batch_import.missing_data_row')
-    end
-
-    def validate_file
-      if @file.nil?
-        raise SampleFileImportError,
-              I18n.t('services.samples.batch_import.empty_file')
-      end
-
-      extension = validate_file_extension
-      download_batch_import_file(extension)
-
-      @headers = @spreadsheet.row(1).compact
-      validate_file_headers
-
-      validate_file_rows
-    end
-
-    def download_batch_import_file(extension)
-      begin
-        @temp_import_file.binmode
-        @file.download do |chunk|
-          @temp_import_file.write(chunk)
-        end
-      ensure
-        @temp_import_file.close
-      end
-      @spreadsheet = if extension.eql? '.tsv'
-                       Roo::CSV.new(@temp_import_file, extension:, csv_options: { col_sep: "\t" })
-                     else
-                       Roo::Spreadsheet.open(@temp_import_file.path, extension:)
-                     end
-    end
+    protected
 
     def perform_file_import # rubocop:disable Metrics/AbcSize,Metrics/MethodLength
       response = {}
@@ -147,11 +57,7 @@ module Samples
       response
     end
 
-    def cleanup_files
-      # delete the blob and temporary file as we no longer require them
-      @file.purge
-      @temp_import_file.unlink
-    end
+    private
 
     def process_sample_row(name, project, description)
       sample_params = { name:, description: }
